requirements.txt
fastapi
uvicorn
streamlit
google-generativeai
chromadb
groq
python-dotenv
requests
pymupdf